Restrict example to 32 bit chromosomes.
git-svn-id: file:///home/mast/dev/jgap/svn/trunk@26 3924feb9-9058-4ba4-b128-58a6af0bd8c6

+++ examples/MaxFunction.java
@@ -35,8 +35,8 @@ public class MaxFunction implements FitnessFunction {
    * as much and, therefore, twice the incentive to select the higher
    * value.
    */
-  public long evaluate(Chromosome subject) {
-    long total = 0;
+  public int evaluate(Chromosome subject) {
+    int total = 0;
     
     for(int i = 0; i < subject.size(); i++) {
       if(subject.getAllele(subject.size() - (i + 1))) {
